requirements.txt
requests==2.22.0
instabot==0.4.5
pillow==7.0.0
python-dotenv==0.11.0